Add files via upload
The electron temperature convention used disregarded that the normalization of potential is different for ions and electrons. A factor alpha that is the fraction of electron mass to ion mass was introduced and multiplies the electric field to get the acceleration in the ionmotion function as well as rescaling the potential for the initialization functions derived using Anders' normalization of the potential (it is scaled to ion mass).